+++ misc/futures-bounded/src/futures_map.rs
